requirements.txt
asgiref==3.8.1
Django==5.0.6
finnhub-python==2.4.20
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
